refactor planner drag and drop
