+++ src/App.tsx
@@ -8,12 +8,12 @@ import {
   CourseEntry,
   CourseType,
 } from "./types/interfaces/Course.interface.ts";
-import { DragDropContext, DropResult } from "@hello-pangea/dnd";
+import { DragDropContext } from "@hello-pangea/dnd";
 import { SemesterType } from "./types/interfaces/Semester.interface.ts";
 import { Filters } from "./types/Filters";
+import { usePlannerDragAndDrop } from "./hooks/usePlannerDragAndDrop";
 
-function App() {
-  // Original state from App.tsx
+export default function App() {
   const [toolboxCourses, setToolboxCourses] = useState<CourseEntry[]>([]);
   const [plannerCourses, setPlannerCourses] = useState<SemesterType[]>([
     {
@@ -24,10 +24,7 @@ function App() {
     },
   ]);
   const [isDragging, setIsDragging] = useState<boolean>(false);
-
   const [searchResults, setSearchResults] = useState<CourseType[]>([]);
-
-  // Moved from SearchBar.tsx
   const [searchPrompt, setSearchPrompt] = useState("");
   const [showFilter, setShowFilter] = useState(false);
   const [filters, setFilters] = useState<Filters>({
@@ -36,236 +33,13 @@ function App() {
     Semesters: [],
   });
 
-  // Reorders a simple array
-  const reorder = <T,>(
-    list: T[],
-    startIndex: number,
-    endIndex: number
-  ): T[] => {
-    const result = [...list];
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
-    return result;
-  };
-
-  const cloneCourse = (course: CourseEntry): CourseEntry => {
-    return {
-      ...course,
-      name: course.name + "-" + Math.random().toString(36).substring(2, 7),
-      data: { ...course.data },
-      count: 1,
-    };
-  };
-
-  const onDragStart = () => {
-    setIsDragging(true);
-  };
-
-  const onDragEnd = (result: DropResult) => {
-    setIsDragging(false);
-    const { source, destination, draggableId } = result;
-    if (!destination) return;
-
-    const sInd = source.droppableId;
-    const dInd = destination.droppableId;
-
-    if (sInd === "toolbox" && dInd === "toolbox") {
-      setToolboxCourses((prev) =>
-        reorder(prev, source.index, destination.index)
-      );
-      return;
-    }
-
-    if (dInd === "garbage" && sInd === "toolbox") {
-      setToolboxCourses((prev) => {
-        const updated = [...prev];
-        updated.splice(source.index, 1);
-        return updated;
-      });
-      return;
-    }
-
-    if (sInd === dInd && dInd.startsWith("planner-")) {
-      const semesterIndex = parseInt(dInd.split("-")[1]);
-      setPlannerCourses((prev) =>
-        prev.map((semester, idx) =>
-          idx === semesterIndex
-            ? {
-                ...semester,
-                courseList: reorder(
-                  semester.courseList,
-                  source.index,
-                  destination.index
-                ),
-              }
-            : semester
-        )
-      );
-      return;
-    }
-
-    if (sInd.startsWith("planner-") && dInd.startsWith("planner-")) {
-      const sourceSemesterIndex = parseInt(sInd.split("-")[1]);
-      const destSemesterIndex = parseInt(dInd.split("-")[1]);
-
-      setPlannerCourses((prev) => {
-        const updated = [...prev];
-        const sourceIdx = sourceSemesterIndex;
-        const destIdx = destSemesterIndex;
-
-        const sourceSemester = updated[sourceIdx];
-        const destSemester = updated[destIdx];
-
-        if (
-          !sourceSemester ||
-          !destSemester ||
-          source.index < 0 ||
-          source.index >= sourceSemester.courseList.length
-        ) {
-          console.warn("Invalid source or destination index");
-          return prev;
-        }
-
-        const courseListCopy = [...sourceSemester.courseList];
-        const [movedCourse] = courseListCopy.splice(source.index, 1);
-
-        if (!movedCourse) {
-          console.warn("Failed to extract course from source");
-          return prev;
-        }
-
-        updated[sourceIdx] = {
-          ...sourceSemester,
-          courseList: courseListCopy,
-          creditsTotal:
-            sourceSemester.creditsTotal - movedCourse.data.credit_max,
-        };
-
-        const destListCopy = [...destSemester.courseList];
-        destListCopy.splice(destination.index, 0, movedCourse);
-
-        updated[destIdx] = {
-          ...destSemester,
-          courseList: destListCopy,
-          creditsTotal: destSemester.creditsTotal + movedCourse.data.credit_max,
-        };
-
-        return updated;
-      });
-
-      return;
-    }
-
-    const fromToolbox = sInd === "toolbox";
-    const toPlanner = dInd.startsWith("planner-");
-
-    if (fromToolbox && toPlanner) {
-      const courseToClone = toolboxCourses.find((c) => c.name === draggableId);
-      const semesterIndex = parseInt(dInd.split("-")[1]);
-
-      if (courseToClone) {
-        const newCourse = cloneCourse(courseToClone);
-
-        setPlannerCourses((prev) =>
-          prev.map((semester, idx) => {
-            if (idx + 1 === semesterIndex) {
-              const updatedList = [...semester.courseList];
-              updatedList.splice(destination.index, 0, newCourse);
-
-              return {
-                ...semester,
-                courseList: updatedList,
-                creditsTotal:
-                  semester.creditsTotal + courseToClone.data.credit_max,
-              };
-            }
-            return semester;
-          })
-        );
-
-        setToolboxCourses((prev) =>
-          prev
-            .map((c) =>
-              c.name === courseToClone.name ? { ...c, count: c.count - 1 } : c
-            )
-            .filter((c) => c.count > 0)
-        );
-      }
-
-      return;
-    }
-
-    // if (sInd.startsWith("planner-") && dInd === "toolbox") {
-    //   const sourceSemesterIndex = parseInt(sInd.split("-")[1]);
-
-    //   setPlannerCourses((prev) => {
-    //     const updated = [...prev];
-    //     const sourceSemester = updated[sourceSemesterIndex - 1];
-    //     const courseListCopy = [...sourceSemester.courseList];
-    //     const [removedCourse] = courseListCopy.splice(source.index, 1);
-
-    //     if (!removedCourse) return prev;
-
-    //     updated[sourceSemesterIndex - 1] = {
-    //       ...sourceSemester,
-    //       courseList: courseListCopy,
-    //       creditsTotal:
-    //         sourceSemester.creditsTotal - removedCourse.data.credit_max,
-    //     };
-
-    //     setToolboxCourses((prevToolbox) => {
-    //       const existing = prevToolbox.find(
-    //         (c) =>
-    //           c.data.dept === removedCourse.data.dept &&
-    //           c.data.code_num === removedCourse.data.code_num
-    //       );
-
-    //       if (existing) {
-    //         return prevToolbox.map((c) =>
-    //           c.name === existing.name ? { ...c, count: c.count + 1 } : c
-    //         );
-    //       } else {
-    //         return [
-    //           ...prevToolbox,
-    //           {
-    //             name: removedCourse.data.dept + removedCourse.data.code_num,
-    //             data: removedCourse.data,
-    //             count: 1,
-    //           },
-    //         ];
-    //       }
-    //     });
-
-    //     return updated;
-    //   });
-
-    //   return;
-    // }
-
-    if (sInd.startsWith("planner-") && dInd === "garbage") {
-      const sourceSemesterIndex = parseInt(sInd.split("-")[1]);
-
-      setPlannerCourses((prev) => {
-        const updated = [...prev];
-        const sourceSemester = updated[sourceSemesterIndex - 1];
-        const courseListCopy = [...sourceSemester.courseList];
-        const [removedCourse] = courseListCopy.splice(source.index, 1);
-
-        if (!removedCourse) return prev;
-
-        updated[sourceSemesterIndex - 1] = {
-          ...sourceSemester,
-          courseList: courseListCopy,
-          creditsTotal:
-            sourceSemester.creditsTotal - removedCourse.data.credit_max,
-        };
-
-        return updated;
-      });
-
-      return;
-    }
-  };
+  // The complex logic is now neatly contained in the hook
+  const { onDragStart, onDragEnd } = usePlannerDragAndDrop({
+    toolboxCourses,
+    setToolboxCourses,
+    setPlannerCourses,
+    setIsDragging,
+  });
 
   return (
     <div className="font-['Helvetica'] min-h-screen">
@@ -335,5 +109,3 @@ function App() {
     </div>
   );
 }
-
-export default App;
